+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @bs-plugins/template
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.7.0
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.6.1
+
 ## 0.0.6
 
 ### Patch Changes